refactor(lib/vector): update references
